Split product/mutations/products.py file
