+++ saleor/graphql/product/mutations/__init__.py
@@ -1,5 +1,59 @@
-from .product_type_create import ProductTypeCreate
-from .product_type_delete import ProductTypeDelete
-from .product_type_update import ProductTypeUpdate
+from .category import CategoryCreate, CategoryDelete, CategoryUpdate
+from .collection import (
+    CollectionAddProducts,
+    CollectionCreate,
+    CollectionDelete,
+    CollectionRemoveProducts,
+    CollectionReorderProducts,
+    CollectionUpdate,
+)
+from .product import (
+    ProductCreate,
+    ProductDelete,
+    ProductMediaCreate,
+    ProductMediaDelete,
+    ProductMediaReorder,
+    ProductMediaUpdate,
+    ProductUpdate,
+)
+from .product_type import ProductTypeCreate, ProductTypeDelete, ProductTypeUpdate
+from .product_variant import (
+    ProductVariantCreate,
+    ProductVariantDelete,
+    ProductVariantPreorderDeactivate,
+    ProductVariantReorder,
+    ProductVariantSetDefault,
+    ProductVariantUpdate,
+    VariantMediaAssign,
+    VariantMediaUnassign,
+)
 
-__all__ = ["ProductTypeCreate", "ProductTypeUpdate", "ProductTypeDelete"]
+__all__ = [
+    "ProductTypeCreate",
+    "ProductTypeUpdate",
+    "ProductTypeDelete",
+    "CategoryCreate",
+    "CategoryDelete",
+    "CategoryUpdate",
+    "CollectionAddProducts",
+    "CollectionCreate",
+    "CollectionDelete",
+    "CollectionRemoveProducts",
+    "CollectionReorderProducts",
+    "CollectionUpdate",
+    "ProductCreate",
+    "ProductDelete",
+    "ProductMediaCreate",
+    "ProductMediaDelete",
+    "ProductMediaReorder",
+    "ProductMediaUpdate",
+    "ProductUpdate",
+    "ProductVariantCreate",
+    "ProductVariantDelete",
+    "ProductVariantPreorderDeactivate",
+    "ProductVariantReorder",
+    "ProductVariantSetDefault",
+    "ProductVariantUpdate",
+    "VariantMediaAssign",
+    "VariantMediaUnassign",
+]
