+++ saleor/graphql/product/mutations/category/__init__.py
@@ -1,0 +1,5 @@
+from .category_create import CategoryCreate
+from .category_delete import CategoryDelete
+from .category_update import CategoryUpdate
+
+__all__ = ["CategoryCreate", "CategoryDelete", "CategoryUpdate"]
