+++ saleor/graphql/product/mutations/collection/__init__.py
@@ -1,0 +1,15 @@
+from .collection_add_products import CollectionAddProducts
+from .collection_create import CollectionCreate
+from .collection_delete import CollectionDelete
+from .collection_remove_products import CollectionRemoveProducts
+from .collection_reorder_products import CollectionReorderProducts
+from .collection_update import CollectionUpdate
+
+__all__ = [
+    "CollectionAddProducts",
+    "CollectionCreate",
+    "CollectionDelete",
+    "CollectionRemoveProducts",
+    "CollectionReorderProducts",
+    "CollectionUpdate",
+]
