+++ saleor/graphql/product/mutations/collection/collection_reorder_products.py
@@ -1,0 +1,96 @@
+import graphene
+from django.core.exceptions import ObjectDoesNotExist, ValidationError
+
+from .....core.permissions import ProductPermissions
+from .....core.tracing import traced_atomic_transaction
+from .....product import models
+from .....product.error_codes import CollectionErrorCode, ProductErrorCode
+from ....channel import ChannelContext
+from ....core.mutations import BaseMutation
+from ....core.types import CollectionError, NonNullList
+from ....core.utils.reordering import perform_reordering
+from ...types import Collection, Product
+
+
+class MoveProductInput(graphene.InputObjectType):
+    product_id = graphene.ID(
+        description="The ID of the product to move.", required=True
+    )
+    sort_order = graphene.Int(
+        description=(
+            "The relative sorting position of the product (from -inf to +inf) "
+            "starting from the first given product's actual position."
+            "1 moves the item one position forward, -1 moves the item one position "
+            "backward, 0 leaves the item unchanged."
+        )
+    )
+
+
+class CollectionReorderProducts(BaseMutation):
+    collection = graphene.Field(
+        Collection, description="Collection from which products are reordered."
+    )
+
+    class Meta:
+        description = "Reorder the products of a collection."
+        permissions = (ProductPermissions.MANAGE_PRODUCTS,)
+        error_type_class = CollectionError
+        error_type_field = "collection_errors"
+
+    class Arguments:
+        collection_id = graphene.Argument(
+            graphene.ID, required=True, description="ID of a collection."
+        )
+        moves = NonNullList(
+            MoveProductInput,
+            required=True,
+            description="The collection products position operations.",
+        )
+
+    @classmethod
+    def perform_mutation(cls, _root, info, collection_id, moves):
+        pk = cls.get_global_id_or_error(
+            collection_id, only_type=Collection, field="collection_id"
+        )
+
+        try:
+            collection = models.Collection.objects.prefetch_related(
+                "collectionproduct"
+            ).get(pk=pk)
+        except ObjectDoesNotExist:
+            raise ValidationError(
+                {
+                    "collection_id": ValidationError(
+                        f"Couldn't resolve to a collection: {collection_id}",
+                        code=ProductErrorCode.NOT_FOUND,
+                    )
+                }
+            )
+
+        m2m_related_field = collection.collectionproduct
+
+        operations = {}
+
+        # Resolve the products
+        for move_info in moves:
+            product_pk = cls.get_global_id_or_error(
+                move_info.product_id, only_type=Product, field="moves"
+            )
+
+            try:
+                m2m_info = m2m_related_field.get(product_id=int(product_pk))
+            except ObjectDoesNotExist:
+                raise ValidationError(
+                    {
+                        "moves": ValidationError(
+                            f"Couldn't resolve to a product: {move_info.product_id}",
+                            code=CollectionErrorCode.NOT_FOUND.value,
+                        )
+                    }
+                )
+            operations[m2m_info.pk] = move_info.sort_order
+
+        with traced_atomic_transaction():
+            perform_reordering(m2m_related_field, operations)
+        collection = ChannelContext(node=collection, channel_slug=None)
+        return CollectionReorderProducts(collection=collection)
