+++ saleor/graphql/product/mutations/product/__init__.py
@@ -1,0 +1,17 @@
+from .product_create import ProductCreate
+from .product_delete import ProductDelete
+from .product_media_create import ProductMediaCreate
+from .product_media_delete import ProductMediaDelete
+from .product_media_reorder import ProductMediaReorder
+from .product_media_update import ProductMediaUpdate
+from .product_update import ProductUpdate
+
+__all__ = [
+    "ProductCreate",
+    "ProductDelete",
+    "ProductMediaCreate",
+    "ProductMediaDelete",
+    "ProductMediaReorder",
+    "ProductMediaUpdate",
+    "ProductUpdate",
+]
