+++ saleor/graphql/product/mutations/product/product_media_create.py
@@ -1,0 +1,125 @@
+import graphene
+import requests
+from django.core.exceptions import ValidationError
+from django.core.files import File
+
+from .....core.permissions import ProductPermissions
+from .....core.utils.validators import get_oembed_data
+from .....product import ProductMediaTypes, models
+from .....product.error_codes import ProductErrorCode
+from ....channel import ChannelContext
+from ....core.mutations import BaseMutation
+from ....core.types import ProductError, Upload
+from ....core.validators.file import (
+    clean_image_file,
+    get_filename_from_url,
+    is_image_url,
+    validate_image_url,
+)
+from ....plugins.dataloaders import load_plugin_manager
+from ...types import Product, ProductMedia
+
+
+class ProductMediaCreateInput(graphene.InputObjectType):
+    alt = graphene.String(description="Alt text for a product media.")
+    image = Upload(
+        required=False, description="Represents an image file in a multipart request."
+    )
+    product = graphene.ID(
+        required=True, description="ID of an product.", name="product"
+    )
+    media_url = graphene.String(
+        required=False, description="Represents an URL to an external media."
+    )
+
+
+class ProductMediaCreate(BaseMutation):
+    product = graphene.Field(Product)
+    media = graphene.Field(ProductMedia)
+
+    class Arguments:
+        input = ProductMediaCreateInput(
+            required=True, description="Fields required to create a product media."
+        )
+
+    class Meta:
+        description = (
+            "Create a media object (image or video URL) associated with product. "
+            "For image, this mutation must be sent as a `multipart` request. "
+            "More detailed specs of the upload format can be found here: "
+            "https://github.com/jaydenseric/graphql-multipart-request-spec"
+        )
+        permissions = (ProductPermissions.MANAGE_PRODUCTS,)
+        error_type_class = ProductError
+        error_type_field = "product_errors"
+
+    @classmethod
+    def validate_input(cls, data):
+        image = data.get("image")
+        media_url = data.get("media_url")
+
+        if not image and not media_url:
+            raise ValidationError(
+                {
+                    "input": ValidationError(
+                        "Image or external URL is required.",
+                        code=ProductErrorCode.REQUIRED,
+                    )
+                }
+            )
+        if image and media_url:
+            raise ValidationError(
+                {
+                    "input": ValidationError(
+                        "Either image or external URL is required.",
+                        code=ProductErrorCode.DUPLICATED_INPUT_ITEM,
+                    )
+                }
+            )
+
+    @classmethod
+    def perform_mutation(cls, _root, info, **data):
+        data = data.get("input")
+        cls.validate_input(data)
+        product = cls.get_node_or_error(
+            info,
+            data["product"],
+            field="product",
+            only_type=Product,
+            qs=models.Product.objects.prefetched_for_webhook(),
+        )
+
+        alt = data.get("alt", "")
+        media_url = data.get("media_url")
+        if img_data := data.get("image"):
+            data["image"] = info.context.FILES.get(img_data)
+            image_data = clean_image_file(data, "image", ProductErrorCode)
+            media = product.media.create(
+                image=image_data, alt=alt, type=ProductMediaTypes.IMAGE
+            )
+        if media_url:
+            # Remote URLs can point to the images or oembed data.
+            # In case of images, file is downloaded. Otherwise we keep only
+            # URL to remote media.
+            if is_image_url(media_url):
+                validate_image_url(media_url, "media_url", ProductErrorCode.INVALID)
+                filename = get_filename_from_url(media_url)
+                image_data = requests.get(media_url, stream=True)
+                image_file = File(image_data.raw, filename)
+                media = product.media.create(
+                    image=image_file,
+                    alt=alt,
+                    type=ProductMediaTypes.IMAGE,
+                )
+            else:
+                oembed_data, media_type = get_oembed_data(media_url, "media_url")
+                media = product.media.create(
+                    external_url=oembed_data["url"],
+                    alt=oembed_data.get("title", alt),
+                    type=media_type,
+                    oembed_data=oembed_data,
+                )
+        manager = load_plugin_manager(info.context)
+        cls.call_event(manager.product_updated, product)
+        product = ChannelContext(node=product, channel_slug=None)
+        return ProductMediaCreate(product=product, media=media)
