+++ saleor/graphql/product/mutations/product_type/__init__.py
@@ -1,0 +1,5 @@
+from .product_type_create import ProductTypeCreate
+from .product_type_delete import ProductTypeDelete
+from .product_type_update import ProductTypeUpdate
+
+__all__ = ["ProductTypeCreate", "ProductTypeDelete", "ProductTypeUpdate"]
