+++ saleor/graphql/product/schema.py
@@ -40,7 +40,35 @@ from .filters import (
     ProductTypeFilterInput,
     ProductVariantFilterInput,
 )
-from .mutations import ProductTypeCreate, ProductTypeDelete, ProductTypeUpdate
+from .mutations import (
+    CategoryCreate,
+    CategoryDelete,
+    CategoryUpdate,
+    CollectionAddProducts,
+    CollectionCreate,
+    CollectionDelete,
+    CollectionRemoveProducts,
+    CollectionReorderProducts,
+    CollectionUpdate,
+    ProductCreate,
+    ProductDelete,
+    ProductMediaCreate,
+    ProductMediaDelete,
+    ProductMediaReorder,
+    ProductMediaUpdate,
+    ProductTypeCreate,
+    ProductTypeDelete,
+    ProductTypeUpdate,
+    ProductUpdate,
+    ProductVariantCreate,
+    ProductVariantDelete,
+    ProductVariantPreorderDeactivate,
+    ProductVariantReorder,
+    ProductVariantSetDefault,
+    ProductVariantUpdate,
+    VariantMediaAssign,
+    VariantMediaUnassign,
+)
 from .mutations.attributes import (
     ProductAttributeAssign,
     ProductAttributeAssignmentUpdate,
@@ -59,32 +87,6 @@ from .mutations.digital_contents import (
     DigitalContentDelete,
     DigitalContentUpdate,
     DigitalContentUrlCreate,
-)
-from .mutations.products import (
-    CategoryCreate,
-    CategoryDelete,
-    CategoryUpdate,
-    CollectionAddProducts,
-    CollectionCreate,
-    CollectionDelete,
-    CollectionRemoveProducts,
-    CollectionReorderProducts,
-    CollectionUpdate,
-    ProductCreate,
-    ProductDelete,
-    ProductMediaCreate,
-    ProductMediaDelete,
-    ProductMediaReorder,
-    ProductMediaUpdate,
-    ProductUpdate,
-    ProductVariantCreate,
-    ProductVariantDelete,
-    ProductVariantPreorderDeactivate,
-    ProductVariantReorder,
-    ProductVariantSetDefault,
-    ProductVariantUpdate,
-    VariantMediaAssign,
-    VariantMediaUnassign,
 )
 from .resolvers import (
     resolve_categories,
