+++ saleor/graphql/product/tests/test_product_discounted_price.py
@@ -7,7 +7,10 @@ from ...discount.enums import DiscountValueTypeEnum
 from ...tests.utils import get_graphql_content
 
 
-@patch("saleor.graphql.product.mutations.products.update_product_discounted_price_task")
+@patch(
+    "saleor.graphql.product.mutations.product_variant.product_variant_delete"
+    ".update_product_discounted_price_task"
+)
 @patch("saleor.order.tasks.recalculate_orders_task.delay")
 def test_product_variant_delete_updates_discounted_price(
     mocked_recalculate_orders_task,
@@ -89,7 +92,7 @@ def test_category_delete_updates_discounted_price(
 
 
 @patch(
-    "saleor.graphql.product.mutations.products"
+    "saleor.graphql.product.mutations.collection.collection_add_products"
     ".update_products_discounted_prices_of_catalogues_task"
 )
 def test_collection_add_products_updates_discounted_price(
@@ -133,8 +136,8 @@ def test_collection_add_products_updates_discounted_price(
 
 
 @patch(
-    "saleor.graphql.product.mutations"
-    ".products.update_products_discounted_prices_of_catalogues_task"
+    "saleor.graphql.product.mutations.collection.collection_remove_products"
+    ".update_products_discounted_prices_of_catalogues_task"
 )
 def test_collection_remove_products_updates_discounted_price(
     mock_update_products_discounted_prices_of_catalogues,
